// mobile/lib/api/v2/types.ts
export interface GalleryCard {
  id: number;
  media_id: string;
  thumbnail: string;
  thumbnail_width: number;
  thumbnail_height: number;
  english_title: string;
  japanese_title: string;
  tag_ids: number[];
  num_pages?: number;
  num_favorites?: number;
  upload_date?: number;
  scanlator?: string;
}